+++ src/main/java/org/fundacionjala/at15/pokemon/commands/SubCommandsCreate.java
@@ -1,0 +1,15 @@
+package org.fundacionjala.at15.pokemon.commands;
+
+import picocli.CommandLine;
+
+import java.util.concurrent.Callable;
+
+@CommandLine.Command(name = ":create", subcommands = {PokemonCommands.class, TrainerCommands.class}, description = "Entry point to create objects in Pokemon world")
+class SubCommandsCreate implements Callable<Integer> {
+
+    @Override
+    public Integer call() {
+        System.out.println("Subcommand :pokemon or :trainer needed");
+        return 0;
+    }
+}
